[IMP] website_profile: set correct 'email_from' in validation email
Currently, the 'email_from' field of the 'validation_email' mail template
is utilizing logged in user's email address. However, in our case, logged
in user is the same as one to which we are sending verification email.
This creates the impression that you are sending the mail to yourself,
which is not correct.

With this commit, we now utilize the receiver's company email address
'email_formatted'. This compute field will first prioritize company's
partner email address, and if not available, it will fallback to the
catchall address of the company. If both are not available, we simply
use logged in user's email as before.

TaskID-2634155

closes odoo/odoo#76380

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>